requirements.txt
requests
psycopg2-binary
asyncpg
nest_asyncio